+++ CHANGELOG.md
@@ -49,7 +49,9 @@ and this project adheres to [Semantic Versioning].
     integration. In case `fd` gets removed it will **NOT** revert to using
     `find`. `fd` and `find` are not CLI or partially CLI compatible.
 
-  - **neovim:** Install `neovim` without any custom configuration.
+  - **neovim:** Install `neovim` with basic configuration (e.g. highlighting
+    current line, relative line numbers, smart case text search, 80 char marker
+    line, use spaces instead of tabs).
 
   - **btop:** Modern system monitor for the CLI. Provided monitoring for disk
     usage, CPU and GPU loads, current processes, network upload and download